cloud_run/ingestion/requirements.txt
google-cloud-bigquery==3.40.*
PyYAML==6.0.*
